requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
chatbotAI==0.3.1.3
colorama==0.4.6
dataclass-factory==2.16
decorator==4.4.2
emoji==2.12.1
Faker==26.0.0
ffmpeg==1.4
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.24.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
loguru==0.7.2
magic-filter==1.0.12
MarkupSafe==2.1.5
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
natsort==8.4.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.22.0
proglog==0.1.10
psutil==6.0.0
pyaes==1.6.1
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seedir==0.5.0
setuptools==71.1.0
shazamio_core==1.0.7
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stockfish==3.28.0
sympy==1.13.1
tenacity==8.5.0
TgCrypto==1.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.4
transformers==4.43.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4
yt-dlp==2024.7.16
